% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseperformances.R
\name{analyseperformances}
\alias{analyseperformances}
\title{Analyse performances}
\usage{
analyseperformances(V_out_target, results)
}
\arguments{
\item{V_out_target}{desired output voltage}

\item{results}{table resulting from the function  rk4buckconverter}
}
\value{
the entire analysis
}
\description{
Analyse performances
}
\examples{
 results <- data.frame(
   times = seq(0, 0.01, length.out = 100),
   i_L = runif(100, min = 0, max = 5),
   V_out = runif(100, min = 0, max = 12)
)
  V_out_target <- 5
analyseperformances(V_out_target, results)
}
