% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{convergece of runge kutta 4}
\usage{
convergence(a, b, c)
}
\arguments{
\item{a}{coefficient of the second derivative}

\item{b}{coefficient of the first derivative}

\item{c}{coefficient of the solution function}
}
\value{
time space maximum for stability and precision
}
\description{
the equation is of the form  a y'' + b y' + c y = f(t)
This function displays the maximum time space for stability and the maximum time space for minimum accuracy
Note: a divergent solution does not mean that Runge Kutta cannot approximate the solution
}
\examples{
convergence(1,1,2)
}
