% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_visualisation.R
\name{current_visualisation}
\alias{current_visualisation}
\title{Current visualisuation}
\usage{
current_visualisation(results)
}
\arguments{
\item{results}{table resulting from the function  rk4buckconverter}
}
\value{
Current visualisation
}
\description{
visualise the output current
}
\examples{
results <- data.frame(
   times = seq(0, 0.01, length.out = 100),
   i_L = runif(100, min = 0, max = 5),
   V_out = runif(100, min = 0, max = 12)
)
current_visualisation(results)
}
