% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duty_cycle.R
\name{duty_cycle}
\alias{duty_cycle}
\title{duty cycle}
\usage{
duty_cycle(V_in, V_out_target)
}
\arguments{
\item{V_in}{input voltage}

\item{V_out_target}{output voltage desired}
}
\value{
duty cycle : The potion's time during the input voltage is active for a period
}
\description{
Evaluate the duty cycle to obtain an output voltage close to the desired voltage
}
\examples{
duty = duty_cycle(12,5)
}
