% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dutycycle.R
\name{dutycycle}
\alias{dutycycle}
\title{duty cycle}
\usage{
dutycycle(V_in, V_out)
}
\arguments{
\item{V_in}{input voltage}

\item{V_out}{output voltage}
}
\value{
duty cycle : The potion's time during the input voltage is active for a period
}
\description{
Evaluate the duty cycle to obtain an output voltage close to the desired voltage
}
\examples{
duty = dutycycle(12,5)
}
