% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_simulation.R
\name{full_simulation}
\alias{full_simulation}
\title{Full simulation}
\usage{
full_simulation(V_in, V_out_target, L, C, R_L, R_charge, t_final, i_L_0, v_C_0)
}
\arguments{
\item{V_in}{input voltage}

\item{V_out_target}{output voltage}

\item{L}{inductance}

\item{C}{capacitor capacity}

\item{R_L}{inductance resistance}

\item{R_charge}{charge resistance}

\item{t_final}{time after which iterations stop}

\item{i_L_0}{initial current of inductance}

\item{v_C_0}{initial voltage of capacitor}
}
\value{
return the graphs and display the analyses
}
\description{
Performs complete analysis and visualization of the DC charger's output current and voltage
}
\examples{
full_simulation ( 12.0, 5.0, 47e-6, 100e-6, 0.1, 2.5, 2e-2, 0, 0 )
}
