% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk4buckconverter.R
\name{rk4buckconverter}
\alias{rk4buckconverter}
\title{buck converter with Runge kutta 4}
\usage{
rk4buckconverter(
  V_in,
  d,
  R_charge,
  R_L,
  L,
  C,
  dt,
  t_final,
  i_L_0 = 0,
  v_C_0 = 0
)
}
\arguments{
\item{V_in}{input voltage}

\item{d}{duty cycle}

\item{R_charge}{charge resistance}

\item{R_L}{inductance resistance}

\item{L}{inductance}

\item{C}{capacitor capacity}

\item{dt}{time space : by how much does the time change at each iteration}

\item{t_final}{time after which iterations stop}

\item{i_L_0}{initial current of inductance}

\item{v_C_0}{initial voltage of capacitor}
}
\value{
a 3 columns table consisting of  times, i_L, V_out
}
\description{
We use Runge Kutta 4 to model the buck converter for simulations
}
\examples{
results <- rk4buckconverter(V_in = 12,
  d = 0.417,R_charge = 2.5,
  R_L = 0.1,L = 47e-6,
  C = 100e-6,
  dt = 1e-5,
  t_final = 0.02
)
}
