% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runge_kutta4.R
\name{runge_kutta4}
\alias{runge_kutta4}
\title{Runge Kutta 4}
\usage{
runge_kutta4(X, A, I, h, t, order)
}
\arguments{
\item{X}{the vector containing the coefficients of the equation X = c(a,b,c) / X =c(b,c)}

\item{A}{the vector constituting the second member A = c(F0,w,p)}

\item{I}{initial condition vector I = c( t0, y0, y'0 ) / I = c( t0, y0 )}

\item{h}{time space : by how much does the time change at each iteration}

\item{t}{the time}

\item{order}{the order of the equation 2 or 1}
}
\value{
a 3 or 2 columns table consisting of  t , derive_y , y  for order 2 / t , y for order 1
}
\description{
the equation is of the form  a y'' + b y' + c y = F0 cos(wt+p)
}
\details{
We use the Runge Kutta 4 method to approximate the solution function
}
\examples{
table2 <- runge_kutta4(c(5,0,10),c(0,0.4,0),c(0,0),0.04,30,2)

table1 <- runge_kutta4(c(5,-10),c(0,0,0),c(0,6),0.1,5,1)
}
