% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{stability of runge kutta 4}
\usage{
stability(a, b, c)
}
\arguments{
\item{a}{coefficient of the second derivative}

\item{b}{coefficient of the first derivative}

\item{c}{coefficient of the solution function}
}
\value{
step size maximum for stability and precision
}
\description{
the equation is of the form  a y'' + b y' + c y = f(t)
This function displays the maximum step size for stability and the maximum step size for minimum accuracy
Note: a divergent solution does not mean that Runge Kutta cannot approximate the solution
}
\examples{
stability(1,1,2)
}
