% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_size.R
\name{step_size}
\alias{step_size}
\title{Step size of charger by Runge Kutta 4}
\usage{
step_size(L, C, R_L, R_charge)
}
\arguments{
\item{L}{inductance}

\item{C}{capacitor capacity}

\item{R_L}{inductance resistance}

\item{R_charge}{charge resistance}
}
\value{
time space maximum for stability and precision
}
\description{
retrieves the characteristics of the charger components and calculates the maximum pitch for stability and precision
}
\examples{
dt_max <- step_size(R_charge = 2.5,R_L = 0.1,L = 47e-6,C = 100e-6)
}
