% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespacecharger.R
\name{timespacecharger}
\alias{timespacecharger}
\title{Time space of charger by Runge Kutta 4}
\usage{
timespacecharger(R_charge, R_L, L, C)
}
\arguments{
\item{R_charge}{charge resistance}

\item{R_L}{inductance resistance}

\item{L}{inductance}

\item{C}{capacitor capacity}
}
\value{
time space maximum for stability and precision
}
\description{
retrieves the characteristics of the charger components and calculates the maximum pitch for stability and precision
}
\examples{
dt_max <- timespacecharger(R_charge = 2.5,R_L = 0.1,L = 47e-6,C = 100e-6)
}
