% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voltage_visualisation.R
\name{voltage_visualisation}
\alias{voltage_visualisation}
\title{Voltage visualisation}
\usage{
voltage_visualisation(results)
}
\arguments{
\item{results}{table resulting from the function  buck_converter_rk4}
}
\value{
Voltage visualisation
}
\description{
visualise the output voltage
}
\examples{
results <- data.frame(
   times = seq(0, 0.01, length.out = 100),
   i_L = runif(100, min = 0, max = 5),
   V_out = runif(100, min = 0, max = 12)
)
voltage_visualisation(results)
}
